pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "jailbreak-arena"
version = "0.1.0"
description = "Adversarial RL security testing for LLM applications. An attacker agent learns to break chatbots while a defender patches the system prompt in real time."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Mithilesh Kumar Lala", email = "your@email.com" }
]
keywords = [
    "llm-security",
    "red-teaming",
    "jailbreak",
    "reinforcement-learning",
    "adversarial-ml",
    "llm-safety",
    "openenv",
    "gymnasium",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
requires-python = ">=3.10"
dependencies = [
    "gymnasium>=0.29.1",
    "groq>=0.13.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.20.0"]
gemini = ["google-genai>=1.0.0"]
azure = ["openai>=1.0.0"]
bedrock = ["boto3>=1.34.0"]
langchain = ["langchain>=0.2.0"]
rl = ["stable-baselines3>=2.3.2"]
all = [
    "openai>=1.0.0",
    "anthropic>=0.20.0",
    "google-genai>=1.0.0",
    "boto3>=1.34.0",
    "langchain>=0.2.0",
    "stable-baselines3>=2.3.2",
]

[project.urls]
Homepage = "https://github.com/Mithilesh-Lala/jailbreak-arena"
Repository = "https://github.com/Mithilesh-Lala/jailbreak-arena"
"Bug Tracker" = "https://github.com/Mithilesh-Lala/jailbreak-arena/issues"
Documentation = "https://github.com/Mithilesh-Lala/jailbreak-arena#readme"
"HuggingFace" = "https://huggingface.co/mkl-01/jailbreak-arena"

[project.scripts]
jailbreak-arena = "jailbreak_arena.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["jailbreak_arena*"]

[tool.setuptools.package-data]
jailbreak_arena = ["*.yaml", "*.yml", "*.json"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]